rename modelselector -> modelproviderselector
